test: finish the harness-test move — sibling imports, Helpers prefix
Completes 67c8d25. The moved files still reached their subjects via
'../helpers/…', which resolves correctly from test/helpers/ but reads as
though they still live a directory up, and both still carried the old
`Setup | ` prefix.
